Fixes for LP64 models; announce IPv6 support.
src/alloc-slab.c (SMALL_CUTOFF): Adjust to a multiple of 8.

src/alloc-srvx.c (srvx_malloc): Fix format string for LP64.

src/md5.h (UINT4): Fix typedef for LP64.

src/opserv.c (cmd_warn): Fix whitespace.

src/proto-p10.c (irc_server): Announce IPv6 support.
git-archimport-id: srvx@srvx.net--2005-srvx/srvx--devo--1.3--patch-33

+++ src/alloc-srvx.c
@@ -63,8 +63,9 @@ srvx_malloc(const char *file, unsigned int line, size_t size)
     assert(block != NULL);
     if (block->magic == ALLOC_MAGIC && block->file_id < file_ids_used) {
         /* Only report the error, due to possible false positives. */
-        log_module(MAIN_LOG, LOG_WARNING, "Detected possible reallocation: %p (called by %s:%u/%u; allocated by %u:%u/%u).",
-                   block, file, line, size, block->file_id, block->line, block->size);
+        log_module(MAIN_LOG, LOG_WARNING, "Detected possible reallocation: %p (called by %s:%u/%lu; allocated by %u:%u/%u).",
+                   block, file, line, (unsigned long)size,
+                   block->file_id, block->line, block->size);
     }
     memset(block, 0, sizeof(*block) + size);
     memcpy((char*)(block + 1) + size, redzone, sizeof(redzone));
